lesson_18

+++ app/controllers/comments_controller.rb
@@ -5,9 +5,11 @@ class CommentsController < ApplicationController
 
   before_action :set_commentable!
   before_action :set_question
+  after_action :verify_authorized
 
   def create
     @comment = @commentable.comments.includes(:user).build(comment_params)
+    authorize(@comment)
 
     if @comment.save
       flash[:success] = t('.success')
@@ -20,6 +22,7 @@ class CommentsController < ApplicationController
 
   def destroy
     comment = @commentable.comments.includes(:user).find(params[:id])
+    authorize(comment)
     comment.destroy
     flash[:success] = t('.success')
     redirect_to question_path(@question)
